+++ op-node/eth/ssz_test.go
@@ -1,4 +1,4 @@
-package l2
+package eth
 
 import (
 	"bytes"
