+++ op-node/testutils/random.go
@@ -1,0 +1,24 @@
+package testutils
+
+import (
+	"math/big"
+	"math/rand"
+
+	"github.com/ethereum/go-ethereum/common"
+)
+
+func RandomHash(rng *rand.Rand) (out common.Hash) {
+	rng.Read(out[:])
+	return
+}
+
+func RandomAddress(rng *rand.Rand) (out common.Address) {
+	rng.Read(out[:])
+	return
+}
+
+func RandomETH(rng *rand.Rand, max int64) *big.Int {
+	x := big.NewInt(rng.Int63n(max))
+	x = new(big.Int).Mul(x, big.NewInt(1e18))
+	return x
+}
